Added guide screen and resources
