+++ pubspec.yaml
@@ -106,6 +106,9 @@ flutter:
     - res/logo/app_logo_light.png
     - res/logo/app_logo_dark.png
     - CHANGELOG.md
+    - res/md/GUIDE.md
+    - res/gif/help_delete_rename.gif
+    - res/gif/help_manual_poll.gif
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
