+++ res/md/GUIDE.md
@@ -1,0 +1,21 @@
+# Welcome to the privacyIDEA authenticator.
+This screen gives you a quick introduction to the app.
+
++ The menu button `⋮` or `…` in the upper right corner of the screen gives access to several things:
+    + Accessing this screen by clicking `Guide`
+    + Manually adding tokens by clicking `Add tokens`
+    + Editing the apps settings by clicking `Settings`
+    + Viewing information about the app by clicking `About`
+
++ Tokens can be added manually as mentioned before or by scanning a QR-Code by clicking the `+` button
+in the lower right corner of the screen.
+
++ Tokens can be deleted and renamed by swiping left and clicking the corresponding button:
+![Renaming and deleting tokens by swiping left](resource:res/gif/help_delete_rename.gif)
+
+# privacyIDEA push token:
+
++ If privacyIDEAs push tokens are used, pending challenges can be polled by swiping down:
+![Manually polling by swiping down](resource:res/gif/help_manual_poll.gif)
+
++ If the app is opened, automatic polling can be activated in the settings.
